requirements.txt
Flask
Flask-WTF
WTForms
Flask-Migrate
Flask-Login
SQLAlchemy
Pillow
python-slugify
gunicorn